+++ examples/actions/daytona_activate_snapshot/action.tf
@@ -1,5 +1,0 @@
-action "daytona_activate_snapshot" "example" {
-  config {
-    id = "snapshot-id"
-  }
-}