+++ examples/actions/daytona_decline_organization_invitation/action.tf
@@ -1,5 +1,0 @@
-action "daytona_decline_organization_invitation" "example" {
-  config {
-    invitation_id = "invitation-id"
-  }
-}