+++ examples/actions/daytona_enroll_sms_mfa/action.tf
@@ -1,3 +1,0 @@
-action "daytona_enroll_sms_mfa" "example" {
-  config {}
-}