+++ examples/actions/daytona_fork_sandbox/action.tf
@@ -1,6 +1,0 @@
-action "daytona_fork_sandbox" "example" {
-  config {
-    sandbox_id_or_name = "sandbox-id-or-name"
-    name               = "forked-sandbox-name"
-  }
-}