+++ examples/actions/daytona_initialize_webhooks/action.tf
@@ -1,5 +1,0 @@
-action "daytona_initialize_webhooks" "example" {
-  config {
-    organization_id = "organization-id"
-  }
-}