+++ examples/actions/daytona_leave_organization/action.tf
@@ -1,5 +1,0 @@
-action "daytona_leave_organization" "example" {
-  config {
-    organization_id = "organization-id"
-  }
-}