+++ examples/actions/daytona_start_sandbox/action.tf
@@ -1,5 +1,0 @@
-action "daytona_start_sandbox" "example" {
-  config {
-    sandbox_id_or_name = "sandbox-id-or-name"
-  }
-}