+++ examples/actions/daytona_stop_sandbox/action.tf
@@ -1,6 +1,0 @@
-action "daytona_stop_sandbox" "example" {
-  config {
-    sandbox_id_or_name = "sandbox-id-or-name"
-    force              = false
-  }
-}