+++ examples/actions/daytona_update_sandbox_last_activity/action.tf
@@ -1,5 +1,0 @@
-action "daytona_update_sandbox_last_activity" "example" {
-  config {
-    sandbox_id = "sandbox-id"
-  }
-}